Unnecessary classes removed
