+++ server/src/main/java/fleet/Fleet.java
@@ -1,5 +1,7 @@
 package fleet;
 
+import model.ShipModel;
+
 import java.util.List;
 
 public interface Fleet {
@@ -12,5 +14,5 @@ public interface Fleet {
 
     int getSize();
 
-    List<Ship> getShips();
+    List<ShipModel> getShips();
 }
