+++ server/src/main/java/fleet/HardcodedFleet.java
@@ -1,5 +1,7 @@
 package fleet;
 
+import model.ShipModel;
+
 import java.util.*;
 
 public class HardcodedFleet implements Fleet {
@@ -7,7 +9,7 @@ public class HardcodedFleet implements Fleet {
     private static final Integer[] HARDCODED_POSITIONS = {0, 1, 2, 3, 17, 18, 19, 34, 40, 41, 44, 47, 54, 57, 61, 77, 78, 82, 85, 99};
     private final List<Integer> fleetPositions;
     private List<Integer> hitFields;
-    private final List<Ship> ships;
+    private final List<ShipModel> ships;
 
 
     public HardcodedFleet() {
@@ -38,7 +40,7 @@ public class HardcodedFleet implements Fleet {
     }
 
     @Override
-    public List<Ship> getShips() {
+    public List<ShipModel> getShips() {
         return new ArrayList<>();
     }
 }
