--- server/src/main/java/fleet/Ship.java
@@ -1,8 +1,0 @@
-package fleet;
-
-import java.util.List;
-
-interface Ship {
-    int getNumberOfMasts();
-    List<Integer> getFields(); //TODO: change Integer to Position
-}
