+++ server/src/test/java/fleet/CustomFleetTest.java
@@ -1,5 +1,7 @@
 package fleet;
 
+import model.ShipModel;
+import model.ShipModelImpl;
 import org.testng.annotations.DataProvider;
 import org.testng.annotations.Test;
 
@@ -26,7 +28,7 @@ public class CustomFleetTest {
         CustomFleet fleet = new CustomFleet();
         int initialFleetSize = fleet.getSize();
         int expectedFleetSize = initialFleetSize + 1;
-        Ship ship = new OneMastShip(Arrays.asList(0));
+        ShipModel ship = new ShipModelImpl(Arrays.asList(0));
 
         fleet.addShip(ship);
 
@@ -43,14 +45,14 @@ public class CustomFleetTest {
     }
 
     @Test(dataProvider = "ships to be added to fleet")
-    public void whenAddShipsAndGetShipsThenReturnShips(Ship[] ships) {
+    public void whenAddShipsAndGetShipsThenReturnShips(ShipModel[] ships) {
         CustomFleet fleet = new CustomFleet();
 
-        for (Ship ship : ships) {
+        for (ShipModel ship : ships) {
             fleet.addShip(ship);
         }
 
-        List<Ship> actualShips = fleet.getShips();
+        List<ShipModel> actualShips = fleet.getShips();
         int fleetSize = fleet.getSize();
 
         assertThat(fleetSize).isEqualTo(ships.length);
@@ -60,23 +62,23 @@ public class CustomFleetTest {
     @Test
     public void givenFleetWithTwoShipsWhenGetAllFleetPositionsThenReturnAllPositions() {
         CustomFleet fleet = new CustomFleet();
-        fleet.addShip(new OneMastShip(Arrays.asList(0)));
-        fleet.addShip(new TwoMastShip(Arrays.asList(1, 2)));
+        fleet.addShip(new ShipModelImpl(Arrays.asList(0)));
+        fleet.addShip(new ShipModelImpl(Arrays.asList(1, 2)));
 
         assertThat(fleet.getFleetPositions()).containsExactly(0, 1, 2);
     }
 
-    private OneMastShip generateOneMastShip() {
+    private ShipModelImpl generateOneMastShip() {
         Random random = new Random();
         Integer[] positions = {random.nextInt(11)};
 
-        return new OneMastShip(new ArrayList<>(Arrays.asList(positions)));
+        return new ShipModelImpl(new ArrayList<>(Arrays.asList(positions)));
     }
 
-    private TwoMastShip generateTwoMastShip() {
+    private ShipModelImpl generateTwoMastShip() {
         Random random = new Random();
         Integer[] positions = {random.nextInt(11), random.nextInt(11)};
 
-        return new TwoMastShip(new ArrayList<>(Arrays.asList(positions)));
+        return new ShipModelImpl(new ArrayList<>(Arrays.asList(positions)));
     }
 }
